XIVY-15944 refactor: add language tool control component
Create the components `LanguageToolControl` and `DeleteLanguage` to prepare for a nicely structured language tool.
