+++ packages/cms-editor/src/main/control/language-tool/LanguageToolControl.tsx
@@ -1,0 +1,15 @@
+import { Flex } from '@axonivy/ui-components';
+import { DeleteLanguage } from './DeleteLanguage';
+
+type LanguageToolControlProps = {
+  deleteSelectedLanguage: () => void;
+  hasSelection: boolean;
+};
+
+export const LanguageToolControl = ({ deleteSelectedLanguage, hasSelection }: LanguageToolControlProps) => {
+  return (
+    <Flex gap={2}>
+      <DeleteLanguage deleteSelectedLanguage={deleteSelectedLanguage} hasSelection={hasSelection} />
+    </Flex>
+  );
+};
